simple_parse() is now parse()
We can directly parse homework text by looking at .stripped_strings,
which simplifies the process and requires only one pass through the html
source. The script searches for a div style, which is going to break if
there is any change in how the homework blocks get displayed. The method
I used to sort the assignments by date is also kind of hacky and
probably could be improved.